// src/api/models/projects-generate-proposal-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ProjectsGenerateProposalCreateRequest
 */
export interface ProjectsGenerateProposalCreateRequest {
    /**
     * ID of the team
     * @type {number}
     * @memberof ProjectsGenerateProposalCreateRequest
     */
    'team_id': number;
    /**
     * List of member IDs to include in the proposal
     * @type {Array<number>}
     * @memberof ProjectsGenerateProposalCreateRequest
     */
    'member_ids': Array<number>;
    /**
     * Project requirements description
     * @type {string}
     * @memberof ProjectsGenerateProposalCreateRequest
     */
    'project_requirements'?: string;
}