// src/api/models/projects-generate-proposal-create200-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ProjectsGenerateProposalCreate200ResponseProposal } from './projects-generate-proposal-create200-response-proposal';
// May contain unused imports in some cases
// @ts-ignore
import type { ProjectsGenerateProposalCreate200ResponseTeamData } from './projects-generate-proposal-create200-response-team-data';
// May contain unused imports in some cases
// @ts-ignore
import type { ProjectsGenerateProposalCreate200ResponseValidation } from './projects-generate-proposal-create200-response-validation';

/**
 * 
 * @export
 * @interface ProjectsGenerateProposalCreate200Response
 */
export interface ProjectsGenerateProposalCreate200Response {
    /**
     * 
     * @type {boolean}
     * @memberof ProjectsGenerateProposalCreate200Response
     */
    'success'?: boolean;
    /**
     * 
     * @type {ProjectsGenerateProposalCreate200ResponseTeamData}
     * @memberof ProjectsGenerateProposalCreate200Response
     */
    'team_data'?: ProjectsGenerateProposalCreate200ResponseTeamData;
    /**
     * 
     * @type {string}
     * @memberof ProjectsGenerateProposalCreate200Response
     */
    'project_requirements'?: string;
    /**
     * 
     * @type {ProjectsGenerateProposalCreate200ResponseProposal}
     * @memberof ProjectsGenerateProposalCreate200Response
     */
    'proposal'?: ProjectsGenerateProposalCreate200ResponseProposal;
    /**
     * 
     * @type {ProjectsGenerateProposalCreate200ResponseValidation}
     * @memberof ProjectsGenerateProposalCreate200Response
     */
    'validation'?: ProjectsGenerateProposalCreate200ResponseValidation;
}